Move game logic into seperate crate
