+++ server/game_logic/src/components/destination.rs
