+++ server/game_logic/src/game_state.rs
@@ -1,5 +1,6 @@
 use std::collections::HashMap;
 
+use legion::Entity;
 use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
@@ -13,3 +14,5 @@ pub struct GameStateCache {
     //TODO Change this to https://docs.rs/chashmap/2.2.2/chashmap/
     pub units: HashMap<String, Unit>,
 }
+
+pub type UidEntityMap = HashMap<String, Entity>;
