+++ server/game_logic/src/resources/time_resource.rs
