--- server/obsolete_Dockerfile
@@ -1,25 +1,0 @@
-FROM rust:1.42 as builder
-WORKDIR /usr/src
-
-# 1a: Prepare for static linking
-RUN apt-get update && \
-    apt-get dist-upgrade -y && \
-    apt-get install -y musl-tools && \
-    rustup target add x86_64-unknown-linux-musl
-
-# 1b: Download and compile Rust dependencies (and store as a separate Docker layer)
-RUN USER=root cargo new myprogram
-WORKDIR /usr/src/myprogram
-COPY Cargo.toml Cargo.lock ./
-RUN cargo build --release --target x86_64-unknown-linux-musl
-
-# 1c: Build the exe using the actual source code
-COPY src ./src
-RUN cargo build --release --target x86_64-unknown-linux-musl
-
-# 2: Copy the exe to an empty Docker image
-FROM scratch
-EXPOSE 8000
-COPY --from=builder /usr/src/myprogram/target/x86_64-unknown-linux-musl/release/learnwebsocket ./learnwebsocket
-USER 1000
-CMD ["./learnwebsocket"]
